+++ changelog.md
@@ -374,4 +374,5 @@ No changes. Erroneous version bump.
 ## Next
 
 * Split itemsTable view helper into 2 files
+* Added actionsTable view helper
 * Added modelLinkList view helper